+++ rs/xnet/payload_builder/src/lib.rs
@@ -330,7 +330,6 @@ impl XNetPayloadBuilderImpl {
         let deterministic_rng_for_testing = Arc::new(None);
         let certified_slice_pool = Arc::new(Mutex::new(CertifiedSlicePool::new(
             Arc::clone(&certified_stream_store),
-            subnet_id,
             metrics_registry,
         )));
         let slice_pool = Box::new(XNetSlicePoolImpl::new(certified_slice_pool.clone()));
@@ -1202,6 +1201,75 @@ pub struct RefillStreamSliceIndices {
     pub byte_limit: usize,
 }
 
+pub fn refill_stream_slice_indices(
+    pool_lock: Arc<Mutex<CertifiedSlicePool>>,
+    own_subnet_id: SubnetId,
+) -> impl Iterator<Item = (SubnetId, RefillStreamSliceIndices)> {
+    let mut result: BTreeMap<SubnetId, RefillStreamSliceIndices> = BTreeMap::new();
+
+    let pool_slice_stats = {
+        let pool = pool_lock.lock().unwrap();
+
+        if pool.byte_size() > POOL_BYTE_SIZE_SOFT_CAP {
+            // Abort if pool is already full.
+            return result.into_iter();
+        }
+
+        pool.peers()
+            // Skip our own subnet, the loopback stream is routed separately.
+            .filter(|&&subnet_id| subnet_id != own_subnet_id)
+            .map(|&subnet_id| (subnet_id, pool.slice_stats(subnet_id)))
+            .collect::<BTreeMap<_, _>>()
+    };
+
+    for (subnet_id, slice_stats) in pool_slice_stats {
+        let (stream_position, messages_begin, msg_count, byte_size) = match slice_stats {
+            // Have a cached stream position.
+            (Some(stream_position), messages_begin, msg_count, byte_size) => {
+                (stream_position, messages_begin, msg_count, byte_size)
+            }
+
+            // No cached stream position, no pooling / refill necessary.
+            (None, _, _, _) => continue,
+        };
+
+        let (witness_begin, msg_begin, slice_byte_limit) = match messages_begin {
+            // Existing pooled stream, pull partial slice and append.
+            Some(messages_begin) if messages_begin == stream_position.message_index => (
+                stream_position.message_index,
+                stream_position.message_index + (msg_count as u64).into(),
+                POOL_SLICE_BYTE_SIZE_MAX.saturating_sub(byte_size),
+            ),
+
+            // No pooled stream, or pooled stream does not begin at cached stream position, pull
+            // complete slice from cached stream position.
+            _ => (
+                stream_position.message_index,
+                stream_position.message_index,
+                POOL_SLICE_BYTE_SIZE_MAX,
+            ),
+        };
+
+        if slice_byte_limit < SLICE_BYTE_SIZE_MIN {
+            // No more space left in the pool for this slice, bail out.
+            continue;
+        }
+
+        result.insert(
+            subnet_id,
+            RefillStreamSliceIndices {
+                witness_begin,
+                msg_begin,
+                // XNetEndpoint only counts message bytes, allow some overhead (measuread: 350
+                // bytes for certification plus base witness, 2% for large payloads).
+                byte_limit: (slice_byte_limit.saturating_sub(350)) * 98 / 100,
+            },
+        );
+    }
+
+    result.into_iter()
+}
+
 /// An async task that refills the slice pool.
 pub struct PoolRefillTask {
     /// A pool of slices, filled in the background by an async task.
@@ -1251,8 +1319,9 @@ impl PoolRefillTask {
 
     /// Queries all subnets for new slices and puts / appends them to the pool after
     /// validation against the given registry version.
-    pub async fn refill_pool(&self, registry_version: RegistryVersion) {
-        let refill_stream_slice_indices = self.pool.lock().unwrap().refill_stream_slice_indices();
+    async fn refill_pool(&self, registry_version: RegistryVersion) {
+        let refill_stream_slice_indices =
+            refill_stream_slice_indices(self.pool.clone(), self.endpoint_resolver.subnet_id);
 
         for (subnet_id, indices) in refill_stream_slice_indices {
             // `XNetEndpoint` URL of a node on `subnet_id`.
